fixed the scripts for building the environment with different number of sixpack servers

+++ reset.sh
@@ -1,6 +1,13 @@
-echo "This resets all environment configuration (INCLUDING SSL CERTIFICATES and HTTPASSWD)"
+echo "This will resets environment configuration"
 echo "Press any to continue or break to cancel"
 read
 rm .buildenv 
 rm -R nginx/build
-rm nginx/certs/*
+
+if [ "$1" == "cert" ]; then
+	echo
+	echo "Are you SURE you want to delete the certificates?"
+	echo "Press any to continue or break to cancel"
+	read
+	rm nginx/certs/*
+fi
